Refactor player bar into two different widgets
Trying to change the layout using breakpoints was starting to become way
too difficult. So split the player bar into 3 different widgets now.
It's a lot more code but conceptually much easier to understand.

Also - layout looks better now.
