+++ src/ui/player_bar/mini_player.rs
@@ -1,4 +1,7 @@
-use crate::{audio::model::AudioModel, ui::player_bar::common::PlayerImp};
+use crate::{
+    audio::model::AudioModel,
+    ui::player_bar::{compact_player_bar::CompactPlayerBar, full_player_bar::FullPlayerBar},
+};
 use adw::prelude::*;
 use glib::Object;
 use gtk::{glib, subclass::prelude::*};
@@ -15,6 +18,14 @@ impl MiniPlayerBar {
         Object::builder().build()
     }
 
+    pub fn full_bar(&self) -> FullPlayerBar {
+        self.imp().full_bar.clone()
+    }
+
+    pub fn compact_bar(&self) -> CompactPlayerBar {
+        self.imp().compact_bar.clone()
+    }
+
     pub fn bind_to_audio_model(&self, audio_model: &AudioModel, bottom_sheet: &adw::BottomSheet) {
         let imp = self.imp();
         if let Err(e) = imp.bottom_sheet.set(bottom_sheet.clone()) {
@@ -22,18 +33,8 @@ impl MiniPlayerBar {
             return;
         }
 
-        if let Err(e) = imp.audio_model.set(audio_model.clone()) {
-            debug!("Audio model already set: {:?}", e);
-        };
-
-        // Bind playback mode menu
-        imp.playback_mode_menu.bind_to_audio_model(audio_model);
-
-        imp.volume_control
-            .bind_property("value", audio_model, "volume")
-            .bidirectional()
-            .sync_create()
-            .build();
+        imp.full_bar.bind_to_audio_model(audio_model);
+        imp.compact_bar.bind_to_audio_model(audio_model);
 
         audio_model.connect_closure(
             "play",
@@ -42,20 +43,7 @@ impl MiniPlayerBar {
                 #[weak(rename_to = player)]
                 self,
                 move |_audio_model: AudioModel| {
-                    player.imp().update_play_pause_button(true);
                     player.reveal();
-                }
-            ),
-        );
-
-        audio_model.connect_closure(
-            "pause",
-            false,
-            glib::closure_local!(
-                #[weak(rename_to = player)]
-                self,
-                move |_audio_model: AudioModel| {
-                    player.imp().update_play_pause_button(false);
                 }
             ),
         );
@@ -67,21 +55,10 @@ impl MiniPlayerBar {
                 #[weak(rename_to = player)]
                 self,
                 move |_audio_model: AudioModel| {
-                    player.imp().update_play_pause_button(false);
                     player.hide();
                 }
             ),
         );
-
-        audio_model
-            .bind_property("position", self, "position")
-            .sync_create()
-            .build();
-
-        audio_model
-            .bind_property("duration", self, "duration")
-            .sync_create()
-            .build();
 
         audio_model.connect_notify_local(
             Some("queue-index"),
@@ -89,8 +66,6 @@ impl MiniPlayerBar {
                 #[weak(rename_to = player)]
                 self,
                 move |audio_model, _| {
-                    player.imp().update_song_info();
-                    // Show player bar when a song is loaded (queue-index >= 0)
                     if audio_model.queue_index() >= 0 {
                         player.reveal();
                     }
@@ -98,11 +73,6 @@ impl MiniPlayerBar {
             ),
         );
 
-        // Initial update
-        self.imp().update_song_info();
-        self.imp().update_play_pause_button(audio_model.playing());
-
-        // Show player bar if there's already a song loaded
         if audio_model.queue_index() >= 0 {
             self.reveal();
         }
@@ -128,73 +98,24 @@ impl Default for MiniPlayerBar {
 }
 
 mod imp {
-    use std::cell::{OnceCell, RefCell};
+    use std::cell::OnceCell;
 
-    use crate::{
-        audio::model::AudioModel,
-        ui::{album_art::AlbumArt, playback_mode::PlaybackModeMenu, player_bar::common::PlayerImp},
+    use crate::ui::player_bar::{
+        compact_player_bar::CompactPlayerBar, full_player_bar::FullPlayerBar,
     };
     use adw::{prelude::*, subclass::prelude::*};
-    use glib::{Properties, WeakRef, subclass::InitializingObject};
-    use gtk::{
-        CompositeTemplate, TemplateChild,
-        glib::{self},
-    };
+    use glib::subclass::InitializingObject;
+    use gtk::{CompositeTemplate, TemplateChild, glib};
 
-    #[derive(CompositeTemplate, Default, Properties)]
+    #[derive(CompositeTemplate, Default)]
     #[template(resource = "/io/m51/Gelly/ui/player_bar/mini_player.ui")]
-    #[properties(wrapper_type = super::MiniPlayerBar)]
     pub struct MiniPlayerBar {
         #[template_child]
-        pub album_art: TemplateChild<AlbumArt>,
+        pub full_bar: TemplateChild<FullPlayerBar>,
         #[template_child]
-        pub title_label: TemplateChild<gtk::Label>,
-        #[template_child]
-        pub artist_button: TemplateChild<gtk::Button>,
-        #[template_child]
-        pub artist_label: TemplateChild<gtk::Label>,
-        #[template_child]
-        pub album_button: TemplateChild<gtk::Button>,
-        #[template_child]
-        pub album_label: TemplateChild<gtk::Label>,
-        #[template_child]
-        pub play_pause_button: TemplateChild<gtk::Button>,
-        #[template_child]
-        pub prev_button: TemplateChild<gtk::Button>,
-        #[template_child]
-        pub next_button: TemplateChild<gtk::Button>,
-        #[template_child]
-        pub position_scale: TemplateChild<gtk::Scale>,
-        #[template_child]
-        pub position_label: TemplateChild<gtk::Label>,
-        #[template_child]
-        pub duration_label: TemplateChild<gtk::Label>,
-        #[template_child]
-        pub scale_position_label: TemplateChild<gtk::Label>,
-        #[template_child]
-        pub scale_duration_label: TemplateChild<gtk::Label>,
-        #[template_child]
-        pub volume_control: TemplateChild<gtk::ScaleButton>,
-        #[template_child]
-        pub info_button: TemplateChild<gtk::Button>,
-        #[template_child]
-        pub lyrics_button: TemplateChild<gtk::Button>,
-        #[template_child]
-        pub favorite_button: TemplateChild<gtk::ToggleButton>,
-        #[template_child]
-        pub playback_mode_menu: TemplateChild<PlaybackModeMenu>,
+        pub compact_bar: TemplateChild<CompactPlayerBar>,
 
-        pub audio_model: OnceCell<AudioModel>,
         pub bottom_sheet: OnceCell<adw::BottomSheet>,
-        pub lyrics_window: RefCell<Option<WeakRef<adw::Window>>>,
-        pub seek_debounce_id: RefCell<Option<glib::SourceId>>,
-        pub favorite_binding: RefCell<Option<glib::Binding>>,
-
-        #[property(get, set)]
-        pub position: RefCell<u32>,
-
-        #[property(get, set)]
-        pub duration: RefCell<u32>,
     }
 
     #[glib::object_subclass]
@@ -204,6 +125,9 @@ mod imp {
         type ParentType = adw::BreakpointBin;
 
         fn class_init(klass: &mut Self::Class) {
+            // pre-registration calls on these types ensure that they are registered before the template reneders
+            FullPlayerBar::static_type();
+            CompactPlayerBar::static_type();
             klass.bind_template();
         }
 
@@ -212,100 +136,12 @@ mod imp {
         }
     }
 
-    #[glib::derived_properties]
     impl ObjectImpl for MiniPlayerBar {
         fn constructed(&self) {
             self.parent_constructed();
-            self.setup_common_signals();
-            self.setup_clickable_labels();
-            self.setup_volume_icons();
         }
     }
 
-    impl adw::subclass::prelude::BreakpointBinImpl for MiniPlayerBar {}
-    impl WidgetImpl for MiniPlayerBar {
-        fn snapshot(&self, snapshot: &gtk::Snapshot) {
-            self.snapshot_background(snapshot);
-            self.parent_snapshot(snapshot);
-        }
-    }
-
-    impl PlayerImp for MiniPlayerBar {
-        fn audio_model(&self) -> &AudioModel {
-            self.audio_model.get().expect("AudioModel not initialized")
-        }
-        fn lyrics_window(&self) -> &RefCell<Option<WeakRef<adw::Window>>> {
-            &self.lyrics_window
-        }
-        fn seek_debounce_id(&self) -> &RefCell<Option<glib::SourceId>> {
-            &self.seek_debounce_id
-        }
-        fn position_storage(&self) -> &RefCell<u32> {
-            &self.position
-        }
-        fn duration_storage(&self) -> &RefCell<u32> {
-            &self.duration
-        }
-        fn play_pause_button(&self) -> &gtk::Button {
-            &self.play_pause_button
-        }
-        fn next_button(&self) -> &gtk::Button {
-            &self.next_button
-        }
-        fn prev_button(&self) -> &gtk::Button {
-            &self.prev_button
-        }
-        fn volume_control(&self) -> &gtk::ScaleButton {
-            &self.volume_control
-        }
-        fn info_button(&self) -> &gtk::Button {
-            &self.info_button
-        }
-        fn position_scale(&self) -> &gtk::Scale {
-            &self.position_scale
-        }
-        fn position_label(&self) -> &gtk::Label {
-            &self.position_label
-        }
-        fn duration_label(&self) -> &gtk::Label {
-            &self.duration_label
-        }
-        fn lyrics_button(&self) -> &gtk::Button {
-            &self.lyrics_button
-        }
-        fn favorite_button(&self) -> &gtk::ToggleButton {
-            &self.favorite_button
-        }
-        fn favorite_binding(&self) -> &RefCell<Option<glib::Binding>> {
-            &self.favorite_binding
-        }
-        fn artist_button(&self) -> &gtk::Button {
-            &self.artist_button
-        }
-        fn album_button(&self) -> &gtk::Button {
-            &self.album_button
-        }
-        fn title_label(&self) -> &gtk::Label {
-            &self.title_label
-        }
-        fn artist_label(&self) -> &gtk::Label {
-            &self.artist_label
-        }
-        fn album_label(&self) -> &gtk::Label {
-            &self.album_label
-        }
-        fn album_art(&self) -> &AlbumArt {
-            &self.album_art
-        }
-
-        fn extra_position_update(&self, position: u32) {
-            self.scale_position_label
-                .set_text(&crate::ui::player_bar::common::format_time(position));
-        }
-
-        fn extra_duration_update(&self, duration: u32) {
-            self.scale_duration_label
-                .set_text(&crate::ui::player_bar::common::format_time(duration));
-        }
-    }
+    impl BreakpointBinImpl for MiniPlayerBar {}
+    impl WidgetImpl for MiniPlayerBar {}
 }
