+++ src/ui/player_bar/mod.rs
@@ -1,3 +1,5 @@
 pub mod big_player;
 pub mod common;
+pub mod compact_player_bar;
+pub mod full_player_bar;
 pub mod mini_player;
